*: copy sysmon and cleanup vendor.
